refactor: move editable-text to proper folder
