+++ public_api.ts
@@ -32,4 +32,4 @@ export * from "./src/app/document-services/match/match.service";
 
 export * from "./src/app/document-services/unique-item/unique-item.service";
 
-export * from "./src/app/editable-text/editable-text.service";
+export * from "./src/app/document-services/editable-text/editable-text.service";
