+++ src/app/bl-connect/bl-connect.module.ts
@@ -6,7 +6,7 @@ import { BranchService } from "../document-services/branch/branch.service";
 import { OpeningHourService } from "../document-services/opening-hour/opening-hour.service";
 import { ApiErrorService } from "../api-error/api-error.service";
 import { UserDetailService } from "../document-services/user-detail/user-detail.service";
-import { EditableTextService } from "../editable-text/editable-text.service";
+import { EditableTextService } from "../document-services/editable-text/editable-text.service";
 import { TokenService } from "../token/token.service";
 import { StorageService } from "../storage/storage.service";
 import { LoginService } from "../login/login.service";
